improve code structure based on review feedback
