+++ .env.template
@@ -33,3 +33,7 @@ LANGCHAIN_TRACING_V2=true
 LANGCHAIN_ENDPOINT="https://api.smith.langchain.com"
 LANGCHAIN_API_KEY=
 LANGCHAIN_PROJECT=
+
+# Optional
+# If the debugger should run the generated code.
+EXECUTE=1
