+++ non_src/tests/manual_tests/projects_files/debugger_scenario_2_files/sample_test.py
@@ -1,0 +1,1 @@
+print("Hello World!")
